Update zelnodev3.sh
updated bootstrap link to have script use the one on apt.zel.cash's server, also have script download update script, and script to deal with the peers.dat changing ownership back to root. DO NOT MERGE til I have submitted PR's for the 2 files to be added to this repo.